+++ azure-storage-common/ChangeLog.md
@@ -1,8 +1,11 @@
 # Change Log azure-storage-common
 
 > See [BreakingChanges](BreakingChanges.md) for a detailed list of API breaks.
-## Version XX.XX.XX:
-- Error message now contains the ErrorCode from the x-ms-error-code header value
+
+## Version 1.1.0:
+
+- Support for 2017-07-29 REST version. Please see our REST API documentation and blogs for information about the related added features.
+- Error message now contains the ErrorCode from the x-ms-error-code header value.
 
 ## Version 1.0.0: